// src/generated/theporndb/types.gen.ts
export type PerformerExtraDataObject = {
  astrology?: string
  birthday?: string
  birthplace?: string
  cupsize?: string
  ethnicity?: string
  eye_colour?: string
  fakeboobs?: boolean
  gender?: string
  haircolor?: string
  height?: string
  measurements?: string
  nationality?: string
  piercings?: string
  tattoos?: string
  weight?: string
}
// ...
export type SceneHashEnum = 'OSHASH' | 'PHASH'
// ...
export type DirectorResource = {
  id?: number
  uuid?: string
  name?: string
  slug?: string
}

/**
 * A resource that contains information about an image
 */
export type MediaResource = {
  id?: number
  url?: string
  size?: number
  order?: number
}

/**
 * Contains information about a performer
 */
export type PerformerResource = {
  id?: string
  _id?: number
  slug?: string
  name?: string
  disambiguation?: string
  bio?: string | null
  rating?: number
  is_parent?: boolean
  extras?: {
    gender?: string | null
    birthday?: string | null
    birthday_timestamp?: number | null
    deathday?: string | null
    deathday_timestamp?: number | null
    birthplace?: string | null
    birthplace_code?: string | null
    astrology?: string | null
    ethnicity?: string | null
    nationality?: string | null
    hair_colour?: string | null
    eye_colour?: string | null
    weight?: string | null
    height?: string | null
    measurements?: string | null
    cupsize?: string | null
    tattoos?: string | null
    piercings?: string | null
    waist?: string | null
    hips?: string | null
    career_start_year?: number | null
    career_end_year?: number | null
    fake_boobs?: boolean
    same_sex_only?: boolean
    links?: {
      [key: string]: string | null
    } | null
  }
  aliases?: Array<string> | null
  image?: string
  thumbnail?: string
  face?: string
  posters?: Array<MediaResource> | null
  site_performers?: Array<PerformerSiteResource> | null
}

export type PerformerSiteResource = {
  /**
   * The performer site uuid
   */
  id?: string
  /**
   * The performer site id
   */
  _id?: number
  /**
   * The performer site id
   */
  site_id?: number
  /**
   * The performer site name
   */
  name?: string
  bio?: string
  /**
   * Whether the performer is a parent entity of another performer
   */
  is_parent?: boolean
  extras?: PerformerExtraDataObject
  image?: string
  thumbnail?: string
  face?: string
  scenes?: Array<SceneResource>
  parent?: PerformerResource
  site?: SiteResource
}

/**
 * Describes a marker start and end time
 */
export type MarkerResource = {
  id?: number
  title?: string
  start_time?: number
  end_time?: number | null
  created_at?: string
}

export type SceneResource = {
  id?: string
  _id?: number
  title?: string
  type?: string
  slug?: string
  external_id?: string
  description?: string
  rating?: number
  site_id?: number
  date?: string
  url?: string
  image?: string
  back_image?: string
  poster_image?: string
  poster?: string
  trailer?: string
  duration?: number
  format?: string
  sku?: string
  background?: {
    full?: string
    large?: string
    medium?: string
    small?: string
  }
  media?: MediaResource
  created?: string
  last_updated?: string
  performers?: Array<PerformerResource>
  site?: SiteResource
  tags?: Array<TagResource>
  hashes?: Array<SceneHashBasicResponse>
  markers?: Array<MarkerResource>
  directors?: Array<DirectorResource>
  /**
   * If this entity is a movie, contains scenes attached
   */
  scenes?: Array<SceneResource>
  /**
   * If this entity is a scene, contains movies this scene appears in
   */
  movies?: Array<SceneResource>
  links?: {
    [key: string]: string | null
  } | null
}

export type SiteResource = {
  uuid?: string
  id?: number
  parent_id?: number
  network_id?: number
  name?: string
  short_name?: string
  url?: string
  description?: string
  rating?: number
  logo?: string
  favicon?: string
  poster?: string
  network?: SiteResource
  parent?: SiteResource
}

export type TagResource = {
  id?: number
  uuid?: string
  name?: string
  parents?: Array<TagResource>
}

/**
 * A resource that contains information about a user
 */
export type UserResource = {
  id?: number
  name?: string
}

export type SceneHashBasicResponse = {
  can_delete?: boolean | null
  created_at?: string
  duration?: number
  hash?: string
  id?: number
  scene_id?: number
  submissions?: number
  type?: SceneHashEnum
  updated_at?: string
}